+++ NEWS.md
@@ -2,6 +2,7 @@ shinyWidgets 0.5.4
 ======================
 
 * New functions: `currencyInput()`, `formatNumericInput()` and `autonumericInput()` to enter numeric value with specified format like a currency, by [@srmatth](https://github.com/srmatth).
+* Ability to update icons in `updateNumericInputIcon()` and `updateTextInputIcon()`.
 
 
 ### Bug fixes